// src/domain/enums/beer-color.ts
export enum BeerColor {
    PALE = "PALE",
    GOLDEN = "GOLDEN",
    WEISS = "WEISS",
    IPA = "IPA",
    SAISON = "SAISON",
    EXTRA_SPECIAL_BITTER = "EXTRA_SPECIAL_BITTER",
    GARDE = "GARDE",
    AMBER = "AMBER",
    DUNKEL = "DUNKEL",
    PORTER = "PORTER",
    STOUT = "STOUT",
    BALTIC_PORTER = "BALTIC_PORTER",
    EXPORT_STOUT = "EXPORT_STOUT",
    IMPERIAL_STOUT = "IMPERIAL_STOUT",
    UNKNOWN = "UNKWOWN",
  }